useStateをchatInput内に移動して再レンダリングされる範囲を入力欄だけに限定した
