+++ frontend/src/features/chat/ChatInput.tsx
@@ -1,19 +1,31 @@
 import { ChangeEvent } from 'react';
+import { socket } from '../../socket';
 
 import { Input } from '../../components/Elements/Input/Input';
 import { SendButton } from '../../components/Elements/Button/SendButton';
 
 import { Container } from '../../components/Layout/Container';
+import { useState } from 'react';
 
-export const ChatInput = ({
-  msg,
-  onChangeAct,
-  sendBtnAct,
-}: {
-  msg: string;
-  onChangeAct: (e: ChangeEvent<HTMLInputElement>) => void;
-  sendBtnAct: () => void;
-}) => {
+export const ChatInput = () => {
+  const [msg, setMsg] = useState('');
+  const onChangeAct = (e: ChangeEvent<HTMLInputElement>) => {
+    // TODO : Shift + Enter で 改行を入れられるようにとかしたい
+    if (e.target.value.slice(-1) === '\n') {
+      if (e.target.value.length === 1) {
+        return;
+      }
+      sendBtnAct();
+      return;
+    }
+    setMsg(e.target.value);
+  };
+
+  const sendBtnAct = () => {
+    socket.emit('message', msg);
+    setMsg('');
+  };
+
   return (
     <>
       <div>
