+++ frontend/src/features/chat/ChatTalkArea.tsx
@@ -40,28 +40,10 @@ const isScrollBottom = (scrollBottomRef: React.RefObject<HTMLDivElement>) => {
 };
 
 export const ChatTalkArea = () => {
-  const [msg, setMsg] = useState('');
   const [chatHistMsgs, setchatHistMsgs] = useState<string[]>([]);
 
   const scrollBottomRef = useRef<HTMLDivElement>(null);
   const [isNeedScroll, setIsNeedScroll] = useState(false);
-
-  const onChangeAct = (e: ChangeEvent<HTMLInputElement>) => {
-    // TODO : Shift + Enter で 改行を入れられるようにとかしたい
-    if (e.target.value.slice(-1) === '\n') {
-      if (e.target.value.length === 1) {
-        return;
-      }
-      sendBtnAct();
-      return;
-    }
-    setMsg(e.target.value);
-  };
-
-  const sendBtnAct = () => {
-    socket.emit('message', msg);
-    setMsg('');
-  };
 
   const onMessage = (data: string) => {
     const needScroll = isScrollBottom(scrollBottomRef);
@@ -87,7 +69,7 @@ export const ChatTalkArea = () => {
         isNeedScroll={isNeedScroll}
         scrollBottomRef={scrollBottomRef}
       />
-      <ChatInput msg={msg} onChangeAct={onChangeAct} sendBtnAct={sendBtnAct} />
+      <ChatInput />
     </Container>
   );
 };
